refactored the client connection request on all database access locations.  A few other goodies for the reviews have also been added.
