+++ server/models/trips/tripModel.js
@@ -1,14 +1,14 @@
 // PostgresQL database Setup
 // Official Documentation: https://github.com/brianc/node-postgres/wiki/pg
 // Guide: http://mherman.org/blog/2015/02/12/postgresql-and-nodejs/
+var helper = require('../../config/helpers.js');
 
 var pg = require('pg');
 // var conString = "postgres://username:password@localhost/database";
-var connectionString = process.env.DATABASE_URL || 'postgres://localhost:5432/test';
 
 var tripTableSure = function () {
 
-  var client = new pg.Client(connectionString);
+  var client = helper.createClient();;
 
   client.connect(function(err) {
     if(err) {
